WIP configurable tracing
This creates a tracing module from parts, and ensures lifecycle methods
are invoked. This is part of getting to where we can inject HttpTracing.

The test needs to be fixed as I don't know best way to configure the
dispatcher. It seems some sharing can be done between play and akka,
possibly.

Help wanted on moving this farther.. once refactored I can help with
getting HttpTracing stitched together.